pyproject.toml
[project]
name = "icebench"
version = "0.1.0"
description = "ICE-Bench is a benchmark specifically designed to assess LLMs for Iterative Code gEneration under progressive requirement refinement scenarios"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bigcodebench",
    "codebleu>=0.7.0",
    "datasets>=4.0.0",
    "docker>=7.1.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-deepinfra>=0.1.0",
    "langchain-deepseek>=0.1.4",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "nltk>=3.9.1",
    "openpyxl>=3.1.5",
    "prettytable>=3.16.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tree-sitter>=0.22.3",
    "tree-sitter-python>=0.23.6",
    "uvicorn>=0.35.0",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
